rename VZVBorrowed to VZVComponents
