+++ utils/zerovec/src/varzerovec/slice.rs
@@ -2,7 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use super::borrowed::VarZeroVecBorrowed;
+use super::components::VarZeroVecBorrowed;
 use super::*;
 use core::marker::PhantomData;
 use core::mem;
